Adding duplicated code example in methods

+++ DuplicatedCode/Refactor/PostController.php
@@ -2,7 +2,7 @@
 
 class PostController extends BaseController 
 {
-    private function isAuthorized()
+    private function authorize()
     {
         if (Auth::guest()) {
             throw new UnauthorizedAccessException;
@@ -12,13 +12,13 @@ class PostController extends BaseController
 
     public function store($data)
     {
-        $this->isAuthorized();
+        $this->authorize();
         $this->repository->save($data);
     }
 
     public function delete($id)
     {
-        $this->isAuthorized();
+        $this->authorize();
         $this->respository->delete($id);
     }
 }
